+++ crates/wasmi/src/instance.rs
@@ -10,7 +10,15 @@ use super::{
     Stored,
     Table,
 };
-use crate::module::{FuncIdx, ModuleImportType};
+use crate::{
+    func::FuncError,
+    module::FuncIdx,
+    Error,
+    ExternType,
+    TypedFunc,
+    WasmParams,
+    WasmResults,
+};
 use alloc::{
     boxed::Box,
     collections::{btree_map, BTreeMap},
@@ -111,27 +119,85 @@ impl InstanceEntity {
     }
 }
 
-/// An iterator over the [`Extern`] declarations of an [`Instance`].
-#[derive(Debug)]
-pub struct ExportsIter<'a> {
-    iter: btree_map::Iter<'a, Box<str>, Extern>,
+/// An exported WebAssembly value.
+///
+/// This type is primarily accessed from the [`Instance::exports`] method
+/// and describes what names and items are exported from a Wasm [`Instance`].
+#[derive(Debug, Clone)]
+pub struct Export<'instance> {
+    /// The name of the exported item.
+    name: &'instance str,
+    /// The definition of the exported item.
+    definition: Extern,
 }
 
-impl<'a> ExportsIter<'a> {
+impl<'instance> Export<'instance> {
+    /// Creates a new [`Export`] with the given `name` and `definition`.
+    pub(crate) fn new(name: &'instance str, definition: Extern) -> Export<'instance> {
+        Self { name, definition }
+    }
+
+    /// Returns the name by which this export is known.
+    pub fn name(&self) -> &'instance str {
+        self.name
+    }
+
+    /// Return the [`ExternType`] of this export.
+    ///
+    /// # Panics
+    ///
+    /// If `ctx` does not own this [`Export`].
+    pub fn ty(&self, ctx: impl AsContext) -> ExternType {
+        self.definition.ty(ctx)
+    }
+
+    /// Consume this [`Export`] and return the underlying [`Extern`].
+    pub fn into_extern(self) -> Extern {
+        self.definition
+    }
+
+    /// Returns the underlying [`Func`], if the [`Export`] is a function or `None` otherwise.
+    pub fn into_func(self) -> Option<Func> {
+        self.definition.into_func()
+    }
+
+    /// Returns the underlying [`Table`], if the [`Export`] is a table or `None` otherwise.
+    pub fn into_table(self) -> Option<Table> {
+        self.definition.into_table()
+    }
+
+    /// Returns the underlying [`Memory`], if the [`Export`] is a linear memory or `None` otherwise.
+    pub fn into_memory(self) -> Option<Memory> {
+        self.definition.into_memory()
+    }
+
+    /// Returns the underlying [`Global`], if the [`Export`] is a global variable or `None` otherwise.
+    pub fn into_global(self) -> Option<Global> {
+        self.definition.into_global()
+    }
+}
+
+/// An iterator over the [`Extern`] declarations of an [`Instance`].
+#[derive(Debug)]
+pub struct ExportsIter<'instance> {
+    iter: btree_map::Iter<'instance, Box<str>, Extern>,
+}
+
+impl<'instance> ExportsIter<'instance> {
     /// Creates a new [`ExportsIter`].
-    fn new(iter: btree_map::Iter<'a, Box<str>, Extern>) -> Self {
+    fn new(iter: btree_map::Iter<'instance, Box<str>, Extern>) -> Self {
         Self { iter }
     }
 
     /// Prepares an item to match the expected iterator `Item` signature.
     #[allow(clippy::borrowed_box)]
-    fn convert_item((name, export): (&'a Box<str>, &'a Extern)) -> (&'a str, &'a Extern) {
-        (&**name, export)
+    fn convert_item((name, export): (&'instance Box<str>, &'instance Extern)) -> Export {
+        Export::new(name, *export)
     }
 }
 
-impl<'a> Iterator for ExportsIter<'a> {
-    type Item = (&'a str, &'a Extern);
+impl<'instance> Iterator for ExportsIter<'instance> {
+    type Item = Export<'instance>;
 
     fn size_hint(&self) -> (usize, Option<usize>) {
         self.iter.size_hint()
@@ -181,17 +247,17 @@ impl InstanceEntityBuilder {
         let mut len_tables = module.len_tables();
         let mut len_memories = module.len_memories();
         for import in module.imports() {
-            match import.item_type() {
-                ModuleImportType::Func(_) => {
+            match import.ty() {
+                ExternType::Func(_) => {
                     len_funcs += 1;
                 }
-                ModuleImportType::Table(_) => {
+                ExternType::Table(_) => {
                     len_tables += 1;
                 }
-                ModuleImportType::Memory(_) => {
+                ExternType::Memory(_) => {
                     len_memories += 1;
                 }
-                ModuleImportType::Global(_) => {
+                ExternType::Global(_) => {
                     len_globals += 1;
                 }
             }
@@ -331,7 +397,17 @@ impl InstanceEntityBuilder {
     }
 }
 
-/// A Wasm module instance reference.
+/// An instantiated WebAssembly [`Module`].
+///
+/// This type represents an instantiation of a [`Module`].
+/// It primarily allows to access its [`exports`](Instance::exports)
+/// to call functions, get or set globals, read or write memory, etc.
+///
+/// When interacting with any Wasm code you will want to create an
+/// [`Instance`] in order to execute anything.
+///
+/// Instances are owned by a [`Store`](crate::Store).
+/// Create new instances using [`Linker::instantiate`](crate::Linker::instantiate).
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 #[repr(transparent)]
 pub struct Instance(Stored<InstanceIdx>);
@@ -358,7 +434,7 @@ impl Instance {
     /// # Panics
     ///
     /// Panics if `store` does not own this [`Instance`].
-    pub(crate) fn get_func(&self, store: impl AsContext, index: u32) -> Option<Func> {
+    pub(crate) fn get_func_by_index(&self, store: impl AsContext, index: u32) -> Option<Func> {
         store
             .as_context()
             .store
@@ -379,6 +455,83 @@ impl Instance {
             .get_export(name)
     }
 
+    /// Looks up an exported [`Func`] value by `name`.
+    ///
+    /// Returns `None` if there was no export named `name`,
+    /// or if there was but it wasn’t a function.
+    ///
+    /// # Panics
+    ///
+    /// If `store` does not own this [`Instance`].
+    pub fn get_func(&self, store: impl AsContext, name: &str) -> Option<Func> {
+        self.get_export(store, name)?.into_func()
+    }
+
+    /// Looks up an exported [`Func`] value by `name`.
+    ///
+    /// Returns `None` if there was no export named `name`,
+    /// or if there was but it wasn’t a function.
+    ///
+    /// # Errors
+    ///
+    /// - If there is no export named `name`.
+    /// - If there is no exported function named `name`.
+    /// - If `Params` or `Results` do not match the exported function type.
+    ///
+    /// # Panics
+    ///
+    /// If `store` does not own this [`Instance`].
+    pub fn get_typed_func<Params, Results>(
+        &self,
+        store: impl AsContext,
+        name: &str,
+    ) -> Result<TypedFunc<Params, Results>, Error>
+    where
+        Params: WasmParams,
+        Results: WasmResults,
+    {
+        self.get_export(&store, name)
+            .and_then(Extern::into_func)
+            .ok_or_else(|| Error::Func(FuncError::ExportedFuncNotFound))?
+            .typed::<Params, Results>(store)
+    }
+
+    /// Looks up an exported [`Global`] value by `name`.
+    ///
+    /// Returns `None` if there was no export named `name`,
+    /// or if there was but it wasn’t a global variable.
+    ///
+    /// # Panics
+    ///
+    /// If `store` does not own this [`Instance`].
+    pub fn get_global(&self, store: impl AsContext, name: &str) -> Option<Global> {
+        self.get_export(store, name)?.into_global()
+    }
+
+    /// Looks up an exported [`Table`] value by `name`.
+    ///
+    /// Returns `None` if there was no export named `name`,
+    /// or if there was but it wasn’t a table.
+    ///
+    /// # Panics
+    ///
+    /// If `store` does not own this [`Instance`].
+    pub fn get_table(&self, store: impl AsContext, name: &str) -> Option<Table> {
+        self.get_export(store, name)?.into_table()
+    }
+
+    /// Looks up an exported [`Memory`] value by `name`.
+    ///
+    /// Returns `None` if there was no export named `name`,
+    /// or if there was but it wasn’t a table.
+    ///
+    /// # Panics
+    ///
+    /// If `store` does not own this [`Instance`].
+    pub fn get_memory(&self, store: impl AsContext, name: &str) -> Option<Memory> {
+        self.get_export(store, name)?.into_memory()
+    }
+
     /// Returns an iterator over the exports of the [`Instance`].
     ///
     /// The order of the yielded exports is not specified.
@@ -386,7 +539,10 @@ impl Instance {
     /// # Panics
     ///
     /// Panics if `store` does not own this [`Instance`].
-    pub fn exports<'a, T: 'a>(&self, store: impl Into<StoreContext<'a, T>>) -> ExportsIter<'a> {
+    pub fn exports<'ctx, T: 'ctx>(
+        &self,
+        store: impl Into<StoreContext<'ctx, T>>,
+    ) -> ExportsIter<'ctx> {
         store.into().store.resolve_instance(*self).exports()
     }
 }