+++ crates/wasmi/src/lib.rs
@@ -41,7 +41,7 @@
 //!     let mut store = Store::new(&engine, 42);
 //!     let host_hello = Func::wrap(&mut store, |caller: Caller<'_, HostState>, param: i32| {
 //!         println!("Got {param} from WebAssembly");
-//!         println!("My host state is: {}", caller.host_data());
+//!         println!("My host state is: {}", caller.data());
 //!     });
 //!
 //!     // In order to create Wasm module instances and link their imports
@@ -56,11 +56,7 @@
 //!     let instance = linker
 //!         .instantiate(&mut store, &module)?
 //!         .start(&mut store)?;
-//!     let hello = instance
-//!         .get_export(&store, "hello")
-//!         .and_then(Extern::into_func)
-//!         .ok_or_else(|| anyhow!("could not find function \"hello\""))?
-//!         .typed::<(), ()>(&mut store)?;
+//!     let hello = instance.get_typed_func::<(), ()>(&store, "hello")?;
 //!
 //!     // And finally we can call the wasm!
 //!     hello.call(&mut store, ())?;
@@ -131,22 +127,20 @@ pub use self::{
         TypedResumableInvocation,
     },
     error::Error,
-    external::Extern,
+    external::{Extern, ExternType},
     func::{Caller, Func, IntoFunc, TypedFunc, WasmParams, WasmResults, WasmRet, WasmType},
     func_type::FuncType,
     global::{Global, GlobalType, Mutability},
-    instance::{ExportsIter, Instance},
+    instance::{Export, ExportsIter, Instance},
     linker::Linker,
     memory::{Memory, MemoryType},
     module::{
-        ExportItem,
-        ExportItemKind,
+        ExportType,
+        ImportType,
         InstancePre,
         Module,
         ModuleError,
         ModuleExportsIter,
-        ModuleImport,
-        ModuleImportType,
         ModuleImportsIter,
         Read,
     },