+++ crates/wasmi/src/module/instantiate/pre.rs
@@ -49,7 +49,7 @@ impl InstancePre {
         if let Some(start_index) = opt_start_index {
             let start_func = self
                 .handle
-                .get_func(&mut context, start_index)
+                .get_func_by_index(&mut context, start_index)
                 .unwrap_or_else(|| {
                     panic!("encountered invalid start function after validation: {start_index}")
                 });